Rename areAllResultsAlreadyPresentOnPnc
