--- packages/core/phase2/lib/generateOperations/areAllResultsAlreadyPresentOnPnc/index.ts
@@ -1,1 +1,0 @@
-export { default as areAllResultsAlreadyPresentOnPnc } from "./areAllResultsAlreadyPresentOnPnc"
