+++ packages/core/phase2/lib/generateOperations/areAllResultsOnPnc/index.ts
@@ -1,0 +1,1 @@
+export { default as areAllResultsOnPnc } from "./areAllResultsOnPnc"
