+++ packages/core/phase2/lib/generateOperations/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/arePncDisposalsMatching.ts
