+++ packages/core/phase2/lib/generateOperations/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/createPncDisposalFromResult/getDisposalTextFromResult/exclusionOrderDisposalText.test.ts
