+++ packages/core/phase2/lib/generateOperations/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/createPncDisposalFromResult/getDisposalTextFromResult/isFailedToAppearWarrantIssued.ts
